chore(components/ProductCard): extract Product card to a global component
