+++ src/components/ProductCard/ProductCard.module.scss
@@ -1,4 +1,4 @@
-@use '../../../../../styles/variables' as *;
+@use '../../styles/variables' as *;
 
 .card {
   max-width: $size-card-width;
