+++ src/components/ProductCard/ProductCard.tsx
@@ -1,7 +1,7 @@
 import { ChevronRight } from 'lucide-react';
-import { Product } from '../../../../../types/products';
-import { Card } from '../../../../../components/Card';
-import { Button } from '../../../../../components/Button';
+import { Product } from '../../types/products';
+import { Card } from '../Card';
+import { Button } from '../Button';
 import { ProductCardThumbnail } from './ProductCardThumbnail';
 import { ProductCardHeader } from './ProductCardHeader';
 import { ProductCardMain } from './ProductCardMain';
@@ -9,13 +9,14 @@ import styles from './ProductCard.module.scss';
 
 type ProductCardProps = {
   product: Product;
+  hasHeader?: boolean;
 };
 
-export const ProductCard = ({ product }: ProductCardProps) => (
+export const ProductCard = ({ product, hasHeader }: ProductCardProps) => (
   <Card className={styles.card}>
     <ProductCardThumbnail thumbnail={product.thumbnail} name={product.name} />
     <Card.Content className={styles.card__content}>
-      <ProductCardHeader brand={product.brand} rating={product.rating} />
+      {hasHeader && <ProductCardHeader brand={product.brand} rating={product.rating} />}
       <ProductCardMain name={product.name} price={product.price} discount={product.discount} />
       <Card.Footer>
         <Button
