+++ src/components/ProductCard/ProductCardHeader/ProductCardHeader.module.scss
@@ -1,4 +1,4 @@
-@use '../../../../../../styles/variables' as *;
+@use '../../../styles/variables' as *;
 
 .product-card-header {
   display: flex;
