+++ src/components/ProductCard/ProductCardMain/ProductCardMain.module.scss
@@ -1,4 +1,4 @@
-@use '../../../../../../styles/variables' as *;
+@use '../../../styles/variables' as *;
 
 .product-card-main {
   display: flex;
