+++ src/components/ProductCard/ProductCardThumbnail/ProductCardThumbnail.module.scss
@@ -1,4 +1,4 @@
-@use '../../../../../../styles/variables' as *;
+@use '../../../styles/variables' as *;
 
 .thumbnail {
   position: relative;
