+++ src/components/ProductCard/ProductCardThumbnail/ProductCardThumbnail.test.tsx
