+++ src/components/ProductCard/ProductCardThumbnail/index.ts
